Converts between Bookmarks and fields correctly on LibreOffice 4.4
The problem is that previously we relied on a LibreOffice behaviour
regarding changing the Range of a deleted field. This used to work but
doesn't work anymore on LibreOffice 4.4 as per:
    https://bugs.documentfoundation.org/show_bug.cgi?id=92420

Now the conversion between bookmarks and fields is re-done in a clearer
way and not relying on the behaviour as we did before.